backend/requirements.txt
# EduERP Backend - Multi-Tenant Education ERP System

## Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

## Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

## Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

## Redis & Caching
redis==5.0.1
aioredis==2.0.1

## Background Tasks
celery==5.3.6

## File Storage
boto3==1.34.14
python-magic==0.4.27

## Email
aiosmtplib==3.0.1
email-validator==2.1.0.post1

## Utilities
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3
orjson==3.9.10

## PDF Generation
reportlab==4.0.8
weasyprint==60.1

## QR Code
qrcode==7.4.2
pillow==10.2.0

## Excel/CSV
pandas==2.1.4
openpyxl==3.1.2

## Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0

## Development
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
